refactor: Fix stationAddEvent Error & Code Modularize
